Always use braces even for single lines.
My thinking on this has evolved a lot since this time,
and my light acceptance of the kernel style has given
way to opposition.

It is more correct to always use explicit braces, and this
helps catch errors if new code is added or unusual code
structures or edge cases emerge.